move `ModAlternate` to `OsuModAlternate` and check if intro has ended
